Refactor validation logic into its own package
